Image sizing tweaks and documentation
This also conditionally disables the autoHeight based on the authored settings of the
image interactive.  Doing that correctly required passing the container reference
into the Runtime component. This might cause a performance problem.

I think the comments in the code document the sizing tweaks. However I think it is
probably best not to merge this because of the complexity and possible performance
issues.

A better solution would be for LARA to notify the interactive when the max viewable
height has changed. Then the interactive can decide if it wants to limit the height it
sends to lara based on this max viewable height. The interactive won't have a way
to adjust its width, but currently this kind of width adjustment would only be visible
if the interactive was wrapped with a teacher edition plugin, so it I don't think the
width adjustment is necessary.

[#173174724]