Track the event of when captions / subtitles change, and track when they are toggled
There are no easy hooks to capture the toggle of a caption so I'm relying on the regular state save interval to check on the current state and report if it has been changed. This keeps us from having to dig through the control bar hierarchy to find the actual toggle.