Custom "alt text" for carousel image buttons
PT Story: https://www.pivotaltracker.com/story/show/176575201

[#176575201]

The carousel allows authors to specify images to be used as buttons in the carousel navbar. The images are displayed by making them the background image for the corresponding buttons. These changes add the option to customize text that appears to screen readers for these image buttons, as well as on mouse over of the buttons. The original PT story calls for "alt text" but since the buttons aren't actually images, we add the text as button text and as title and aria-label attributes.