Carousel subquestion response issues
PT Story: https://www.pivotaltracker.com/story/show/176001600

[#176001600]

In some cases, prompts for fill-in-the-blank questions in a carousel were appearing in the response area of other questions in the carousel. These changes set the `subState` value within the iteration through each subinteractive. That ensures the subState is correctly set and then used for each interactive's state.