Add Y-Axis label; default to dataset's label
By default, the graph will show the label of the first dataset. E.g.
for Drag and Drop this will be "Value." Not sure what should happen if
there are multiple datasets with different labels, or if this is even
allowed.

If the user unchecked this, they can either leave the Y-Axis label blank
or put in a custom one.