Add Side-by-Side interactive to nest two q's next to each other
This is architecturally almost identical to the carousel interactive,
but presents itself as two nested question interactives side by side in
iframes.

Add Drag-and-Drop and Graphs as interactive that can be added.

This will cause Drag-and-Drop and Graphs to show up in Carousel and
Scaffolded Questions authoring drop-downs. While the list of embeddable
library interactives could include a filtering key, it seems that we can
also just rely on authors to only embed interactives that make sense.