Allow author to set Side-by-Side data as linked interactive
While nested interactives can't link to other top-level interactives to
consume their data, we will want the graph (or other consumer) to link
to its sibling interactive.

This allows us to specify it blindly, without needing to have a specific
id of the other interactive, since the Side-by-Side runtime will be
handling the data connection anyway.