Include local linked interactive in linkedInteractives array
Because of the way the `useLinkedInteractives` hook works, if we specify
a linked interactive in authoring but it is not included in the
`linkedInteractives` array on `init`, we will get that linked interactive
blown away.

This tries to handle that in as generic a manner as possible. It does
need to know the name of the prop that might have a linked interactive
specified, though (`dataSourceInteractive1` for graph).

I added this to the library interactive array used by the container
interactives, but it could perhaps go elsewhere.