Pass data changes between linked local interactives
If one of the child interactives contains a linked interactive, it will
call `addLinkedInteractiveStateListener`. This would normally be ignored
by the iframe-runtime in Carousel and Scaffolded Questions, but if we
pass in a `addLocalLinkedDataListener` to the iframe-runtime, it will
call that with the request and passes an iframePhone connection.

The Side-by-Side runtime adds this listener. If it is used, every time
a child's `dataset` property changes, it posts that data to that 
iframePhone's connection.