Maintain dragged position in target
This changes the Drag and Drop UI such that dragged items stay where they
were dropped, instead of snapping to the top-left corner.

This is done by rendering all draggable objects in the main container
(instead of having the targets render their own objects, which was causing
lots of excess positioning calculations), and updating the item positions
when an item is dropped in a target.