Support interactiveApi with the full-screen Scaling interactive [#181103123]
This change upgrades the lara client client to the current version and passes the full initInteractive message passed by AP to the fullscreen interactive mixed in with the current initMessage the fullscreen interactive creates.  The interactiveApi in CFM needs the info in the full message.

It was decided in this PR to only address the fullscreen interactive.  This same init message passing may need to be done for the other interactives that use the iframe-runtime container.