Fix drag-and-drop interactive does not look correct in report/dashboard.
PT Story: https://www.pivotaltracker.com/story/show/181154848

[#181154848]

When the interactive is being displayed within a teacher dashboard or report, it will be scaled to fit the confines of its container and not display its prompt. If the optional `view` URL param is present and set to "standalone", the interactive should not be scaled and its prompt should be displayed.