Record open response runtime -- autosave on nav
PT Story: https://www.pivotaltracker.com/story/show/181223914

[#181223914]

These changes fix a problem where an audio file was not properly saved if the user navigated away from the active page before the recording process was allowed to complete. An audio file should be saved and attached to the question/response even if a user starts the recording process and then navigates to another page without either stopping the recording, or allowing it to fully save after stopping the recording.