Align Lato font link with AP [#180588632]
AP uses 400, 700, and 900 font-weight for Lato.  This change makes question-interactives aligns with that plus adds 100 as locked-info.scss uses lighter for a font-weight which is 100 when at normal 400 weight.

This also sets the $default-font-weight SCSS variable to 400.  Currently this variable is not used but may be used in the future.