fix: Authoring form styling after font size change to rems [PT-185043384]
This is a sledgehammer approach to fixing the issue of the authoring form styling breaking after the bootstrap file was converted to scss and the fonts were changed to rems from px to accomodate the font-size feature.

This fix adds a second copy of the original unmodified bootstrap.css file within a :global section of the authoring.scss file.  This locallizes the original bootstrap layout to the authoring form area ensuring it lays out the same as it did before the font size update.