fix: Changed iframe runtime to accept readOnly property [PT-184981533]
This fixes an issue with the font-size of scafolded questions after they are submitted as it was then showing the iframe in report mode which doesn't support the font-size change.

To set the iframe as read only pointer events are disabled and the tab index is removed to prevent tabbing into the iframe.

The carousel interactive was updated to use this new prop instead of adding a disabled div around the iframe.