fix: Default answer disappearing in authoring [PT-185100790]
The default answer was disappearing in open response authoring when the required checkbox was selected as it was not specified in the UI order AND the authoring iframe useAutoHeight() hook was not updating the height when the prediction feeback field showed after required was selected.

This change sets the UI order and makes the prediction feedback always show at the bottom of the form so it is not cut-off from view.