fix: Hide submit button when draw tool dialog showing [PT-188398397]
The submit button, rendered in the BaseQuestionApp component, was showing at the bottom of the draw tool dialog in the image question app.

This fix centralizes the url param code that is used to construct the url show in the drawing tool dialog and then sets the existing (optional) disableSubmitBtnRendering prop to the BaseQuestionApp to hide the submit button when the current url has the drawingToolDialog url param present.