feat: Add demo object storage support [QI-77]
This adds a objectStorageConfig type of "demo" to the iframe-runtime used by the demo system.  It then imports the ObjectStorageProvider in the base runtime that uses that config.

THIS SHOULD NOT BE MERGED UNTIL AP IS UPDATED TO SEND THE OBJECT STORE CONFIG AND THE FINAL VERSIONS OF THE DEPENDENCIES HAVE BEEN PUBLISHED.