feat: Add authorable max recording time to agent simulation interactive [QI-94]
This commit introduces a new authorable property, "Maximum Recording Time",
to the Agent Simulation interactive. This property allows authors to set a
limit on the duration of recordings made during the simulation.

The maximum recording time is be specified in seconds, with a minimum of
1 second and a maximum of 300 seconds (5 minutes). The interactive enforces
this limit by checking the setting to ensure it is in the bounds of 1 to
300 seconds.

Existing interactives will default to a maximum recording time of 90 seconds
to maintain current behavior.  This is done in the runtime component by
spreading the default authored state, which has a maxRecordingTime of 90,
with the authored state received from props.